refactor(project): organize project folders and files
